improvement - allow partial head and footer for suggestions
previously the header and footer for suggestions had to be full html elements. 
Example:    head=<div>The Head</div> and footer=<div>The footer</div>. 
If we wanted to have:
header=<ul>
footer=</ul>
and have the suggestions to be <li>'s, the typeahead wouldn't work. 
This change makes it so that it works in both scenarios.